Time: 0 ms (100%), Space: 10.5 MB (6.15%) - LeetHub

+++ 3754-concatenate-non-zero-digits-and-multiply-by-sum-i/3754-concatenate-non-zero-digits-and-multiply-by-sum-i.cpp
@@ -16,7 +16,7 @@ public:
         }
         return x;
     }
-    int sumOfDigits(int n) {
+    int sumOfDigits(long long n) {
         int sum = 0;
         while(n > 0) {
             int rem = n%10;
